+++ pkgs/native_assets_builder/test/build_runner/link_dry_run_test.dart
@@ -74,10 +74,11 @@ void main() async {
           linkingEnabled: true,
         );
         expect(buildResult.success, true);
-        expect(_getNames(buildResult.assets), orderedEquals(builtHelperAssets));
+        expect(
+            _getNames(buildResult.assets), unorderedEquals(builtHelperAssets));
         expect(
           _getNames(buildResult.assetsForLinking['complex_link']!),
-          orderedEquals(assetsForLinking),
+          unorderedEquals(assetsForLinking),
         );
 
         final linkResult = await linkDryRun(
@@ -88,7 +89,7 @@ void main() async {
         );
         expect(linkResult.success, true);
 
-        expect(_getNames(linkResult.assets), orderedEquals(linkedAssets));
+        expect(_getNames(linkResult.assets), unorderedEquals(linkedAssets));
       });
     },
   );