+++ pkgs/native_assets_builder/test_data/native_add_add_source/hook/build.dart
@@ -6,10 +6,9 @@ import 'package:logging/logging.dart';
 import 'package:native_assets_cli/native_assets_cli.dart';
 import 'package:native_toolchain_c/native_toolchain_c.dart';
 
-const packageName = 'native_add';
-
 void main(List<String> arguments) async {
   await build(arguments, (config, output) async {
+    final packageName = config.packageName;
     final cbuilder = CBuilder.library(
       name: packageName,
       assetName: '${packageName}_bindings_generated.dart',