+++ pkgs/native_assets_cli/lib/src/api/build.dart
@@ -26,7 +26,6 @@ import 'build_output.dart';
 ///       sources: [
 ///         'src/$packageName.c',
 ///       ],
-///       dartBuildFiles: ['hook/build.dart'],
 ///     );
 ///     await cbuilder.run(
 ///       buildConfig: config,
@@ -67,7 +66,6 @@ import 'build_output.dart';
 ///
 ///       output.addDependencies([
 ///         assetSourcePath,
-///         config.packageRoot.resolve('hook/build.dart'),
 ///       ]);
 ///     }
 ///